requirements.txt
Django~=4.1.4
django-bootstrap4
django-allauth
django-simple-menu